Implement Food Transaction

+++ model/food_trx_response.go
@@ -1,8 +1,8 @@
 package model
 
 type FoodTrxResponse struct {
-	Id       int    `json:"id"`
+	Id       int64  `json:"id"`
 	Name     string `json:"name"`
-	SellerId int    `json:"seller_id"`
+	SellerId int64  `json:"seller_id"`
 	Seller   string `json:"seller"`
 }
